Move Environment to new module
